64bit-friendly pdc.h and iomod.h and correspondent changes elsewhere

+++ sys/arch/hppa/stand/libsa/itecons.c
@@ -1,4 +1,4 @@
-/*	$OpenBSD: itecons.c,v 1.7 2003/04/16 07:20:50 mickey Exp $	*/
+/*	$OpenBSD: itecons.c,v 1.8 2004/04/07 18:24:20 mickey Exp $	*/
 
 /*
  * Copyright (c) 1998-2004 Michael Shalayeff
@@ -133,8 +133,8 @@ ite_init(cn)
 		      sizeof(struct pz_device));
 	}
 
-	cons_pzdev->pz_iodc_io = cniodc;
-	kbd_pzdev->pz_iodc_io = kyiodc;
+	cons_pzdev->pz_iodc_io = (u_int)cniodc;
+	kbd_pzdev->pz_iodc_io = (u_int)kyiodc;
 #ifdef DEBUG
 	if (!kyiodc)
 		printf("ite_init: no kbd\n");
